--- a/deeplake/api/dataset.py
+++ b/deeplake/api/dataset.py
@@ -386,6 +386,7 @@ class dataset:
             verbose (bool): If True, logs will be printed. Defaults to True.
             lock_timeout (int): Number of seconds to wait before throwing a LockException. If None, wait indefinitely
             lock_enabled (bool): If true, the dataset manages a write lock. NOTE: Only set to False if you are managing concurrent access externally.
+
         Returns:
             Dataset: Dataset created using the arguments provided.