… deeplake/core/vectorstore/vector_search/dataset/dataset.py
… deeplake/core/vectorstore/vector_search/dataset/test_dataset.py
… deeplake/core/vectorstore/vector_search/utils.py
… deeplake/util/bugout_reporter.py