## 🚀 🚀 Pull Request

### Impact

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes expected existing functionality)
- [ ] Enhancement/New feature (adds functionality without impacting existing logic)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)


### Description
In this PR I am doing 2 main changes:
1. Enabling running indra locally if either user is logged in or have specified token during vectorstore init
2. Removing duplicate backend calls from bugout reporter. Previously when doing VectorStore init, bugout was sending 2 backend requests, one during actual VectorStore init another when the dataset creation/loading logic was called on a lower level. Now we are not doing backend requests to fetch a username, instead we just decode it directly from token that was either specified as token, exposed to env or user has logged in using CLI. This resulted in 2s reduction during dataset creation/loading:
3. If you're logged in use `compute_engine`, while if you log out automatically change to python 

Some Benchmarking on dataset creation using VecotrStore:

| type of dataset                    | Before the change (s)     | After the change (s)       |
|-----------------------------------|---------------------------|----------------------------|
| local datasets without token      | 0.022420883178710938      | 0.037683963775634766       |
| local datasets with token         | 2.718925952911377         | 0.037683963775634766       |
| hub datasets with indra           | 16.12098789215088         | 12.905412197113037         |
| hub datasets with tensor_db       | 15.676331996917725        | 12.73162317276001          |

<!--
A clear and concise description of the change being made.  

The title should introduce what was/will be done
  - Titles show in release notes and search results, so make them useful 
  - Be specific about what was fixed or changed
  - Imagine yourself looking for this PR in 6 months -- what will make it findable and valuable to future you?
  - Good Example: `Correctly apply passed S3 credentials when using VectorStore`
  - Bad Example: `Fixed creds issue`  

The description gives the details on what was/will be done 
- If there is an existing issue this addresses, include "Fixes #XXXX" to auto-link the issue to this PR
- If there is NOT an existing issue, consider creating one.
  - In general, issues describe wanted change from an end-user perspective and PRs describe the technical change.
  - If this change is very small and not worth splitting off an issue, include `Steps To Reproduce`, `Expected Behavior`, and `Actual Behavior` sections in this PR as you would have in the issue.
- Describe what users need to know and how the fix will affect them
- Describe how the code change addresses the problem
- Imagine yourself looking for this PR in 6 months -- what will make it findable and valuable to future you?
- Ensure private information is redacted.
-->

### Things to be aware of
May be its worth testing thoroughly bugout username reporting with different configurations
<!--
- Describe the technical choices you made
- Describe impacts on the codebase
-->

### Things to worry about

<!--
- List any questions or concerns you have with the change
- List unknowns you have 
-->

### Additional Context

<!--
Add any other context about the problem here.
-->